Completing the implementation of the group details data fetching

+++ src/main/java/com/education/education/groups/DTO/response/GroupMessageActivityDataRes.java
@@ -4,5 +4,5 @@ import java.time.LocalDate;
 
 public record GroupMessageActivityDataRes(
     LocalDate date,
-    Integer messageCount) {
+    Long messageCount) {
 }
